+++ gcloud-resize/models.py
