+++ gcloud-resize/utils.py
